audio-ril-interface: fix hwrouting on GTA04A4
It seems that on GTA04 A4, microphone voice routing to
the modem only works when a capture channel is active.

Thus we need to open the mic input PCM and read a single
sample, in order to have the capture channel open during
the call and make the mic work.